requirements.txt
python-dotenv
fastapi[standard]
uvicorn
python-jose[cryptography]
passlib
bcrypt
aiomysql
tortoise-orm[aiomysql]
aerich
tomlkit
slowapi
mangum
sentry-sdk[fastapi]
langchain
langchain-openai
langchain-openrouter
langgraph-checkpoint-postgres
psycopg[binary]
langgraph-cli[inmem]

ruff